ZEN-10085 Rename files to match eslint rules
